refactor(cli): extract per-target search-input derivation into a tested seam
Lift the pure functions that derive a search's inputs from the target —
the register pool, immediate pool, and live-out contract — out of the
6,417-line binary crate into a new `search_inputs` library module, the
search-input analogue of `candidate_windows` (#744).

These functions encode soundness-critical live-out policy (a held-fixed
terminator vetoes register narrowing, ADR-0006/ADR-0008) and search-space
policy (the historical X0..X7 scalar pool; RSP/RBP-excluded x86 pools;
enumerative immediate seeding), yet were reachable only through the
binary's own cli_helper_tests. Moving them to a public library seam makes
every rule a fixture-free unit test and shrinks main.rs by ~490 lines.

- New module src/search_inputs.rs: aarch64_search_registers,
  aarch64_search_immediates (extracted from an inline literal),
  x86_registers_from_target, x86_enumerative_immediates_from_target,
  live_out_for_optimization_prefix, x86_live_out_for_optimization.
- 8 new test-first unit tests cover the three pool functions (previously
  untested directly) plus the new AArch64 immediate-pool helper.
- 11 existing live-out / register-pool regressions relocated intact from
  main.rs (ADR-0006/ADR-0008 soundness gates preserved).
- main.rs imports the seam; runtime behavior is unchanged.
