+++ src/lib.rs
@@ -10,6 +10,7 @@ pub mod isa;
 pub mod parser;
 pub mod report;
 pub mod search;
+pub mod search_inputs;
 pub mod semantics;
 pub mod validation;
